Refactor the tryExecute loop
The major change this provides is removal of `onSpinWait` from
spin-miss paths because in these cases there's nothing to wait for:
a race has already been lost.

Secondly, this reduces duplicate code in the spin-miss path
and consolidates spin-miss code to the end of the loop after
conditionals have been chekced without returning.